requirements.txt
tinyrecord==0.2.0
paho_mqtt==1.5.1
pycryptodome==3.9.9
tinydb==4.3.0
